DEV - CRUD ok expepted missions

+++ Router/contactRouter.php
@@ -12,10 +12,14 @@ if (isset($_GET['action'])) {
                 echo 'Erreur : aucun identifiant de cible envoyé';
             }           
         break;
-        
+    
         case 'edit' : 
-            if (isset($_GET['id']) && $_GET['id'] > 0) {               
+            if (isset($_GET['id']) && $_GET['id'] > 0) {
+                if (isset($_SESSION["ADMIN"]) && $_SESSION["ADMIN"] == "yes") {                
                     editContact($_GET['id']);
+                } else { 
+                    echo 'Erreur : Vous n\'avez pas les droits';  
+                }    
             } else {
                 echo 'Erreur : aucun identifiant de cible envoyé';
             }           
@@ -26,9 +30,13 @@ if (isset($_GET['action'])) {
         break;
         
         case 'delete' :         
-            if (isset($_GET['id']) && $_GET['id'] > 0) {               
-                deleteContact($_GET['id']);
-        }   else {
+            if (isset($_GET['id']) && $_GET['id'] > 0) {
+                if (isset($_SESSION["ADMIN"]) && $_SESSION["ADMIN"] == "yes") {                
+                    deleteContact($_GET['id']);
+                } else { 
+                    echo 'Erreur : Vous n\'avez pas les droits';  
+                } 
+            } else {
             echo 'Erreur : aucun identifiant de cible envoyé';
         }               
         break;
